Azure_bulk/requirements.txt
azure-mgmt-resource==21.1.0
azure-mgmt-compute==1.3.1
azure-mgmt-network==20.0.0
azure-identity==